Resolve unsigned/signed mismatch warnings.
- Working towards issue #11.

+++ src/encmain.c
@@ -353,7 +353,7 @@ int main(int argc, char *argv[])
     // TODO: add more than one reference
 
     // Remove the ref pic (if present)
-    if (encoder->ref->used_size == encoder->cfg->ref_frames) {
+    if (encoder->ref->used_size == (uint32_t)encoder->cfg->ref_frames) {
       picture_list_rem(encoder->ref, encoder->ref->used_size-1, 1);
     }
     // Add current picture as reference
